refactor: simply common add schematic
With this change, we don't export an actual schematic but rather we export a common rule.

This will simply schema changes as we don't have to keep 6 schemas in sync (json and ts)
