+++ modules/common/schematics/testing/test-app.ts
@@ -11,8 +11,7 @@ import {Observable} from 'rxjs';
 import {switchMap} from 'rxjs/operators';
 
 /** Path to the collection file for the NgUniversal schematics */
-export const collectionPath =
-  require.resolve('nguniversal/modules/common/schematics/collection.json');
+export const collectionPath = require.resolve('@schematics/angular/collection.json');
 
 /** Create a base app used for testing. */
 export function createTestApp(appOptions = {}): Observable<UnitTestTree> {
